requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.0
aiosignal==1.4.0
annotated-doc==0.0.2
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.14.2
bitarray==3.7.2
bitsandbytes==0.45.3
bitsandbytes-windows==0.37.5
black==23.11.0
blinker==1.9.0
build==1.3.0
cachelib==0.13.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.2.2
click==8.3.0
cmake==4.1.2
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.11.0
cryptography==41.0.7
cytoolz==1.1.0
dataclasses-json==0.6.7
datasets==4.3.0
defusedxml==0.7.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durationpy==0.10
email-validator==2.1.0
eth-account==0.10.0
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
faiss-cpu==1.12.0
fastapi==0.120.0
filelock==3.20.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.5.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
googlesearch-python==1.3.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.71.2
gspread==6.2.1
gunicorn==21.2.0
h11==0.16.0
hexbytes==0.3.1
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.0.2
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
llama_cpp_python==0.3.16
lru-dict==1.2.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.19.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
mss==10.1.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.6.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parsimonious==0.10.0
pathspec==0.12.1
peft==0.17.1
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.1
py-ecc==8.0.0
py-solc==3.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
pyunormalize==17.0.0
pywin32==311
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2025.10.23
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rlp==4.1.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
semantic-version==2.10.0
sentence-transformers==5.1.2
sentry-sdk==1.38.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.2
starlette==0.48.0
starlette-context==0.3.6
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toolz==1.1.0
torch==2.10.0.dev20251025+cu128
torchaudio==2.10.0.dev20251024+cu128
torchvision==0.25.0.dev20251024+cu128
tqdm==4.67.1
transformers==4.57.1
trl==0.24.0
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.38.0
vertexai==1.71.1
watchfiles==1.1.1
wcwidth==0.2.14
web3==6.20.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
youtube-transcript-api==1.2.3
yt-dlp==2025.10.22
zipp==3.23.0
zstandard==0.25.0
